*** empty log message ***

+++ src/base/objetoJogo.cxx
@@ -103,22 +103,22 @@ void ObjetoJogo::set_setor(PT(Setor) setor){
 }
 
 /*! Retorna valor nutricional do ObjetoJogo */
-int ObjetoJogo::get_valor_nutricional(){
+float ObjetoJogo::get_valor_nutricional(){
 	return valor_nutricional;
 }
 
 /*! Define valor nutricional do ObjetoJogo */
-void ObjetoJogo::set_valor_nutricional(int novo_valor_nutricional){
+void ObjetoJogo::set_valor_nutricional(float novo_valor_nutricional){
 	valor_nutricional = novo_valor_nutricional;
 }
 
 /*! Retorna valor de hidratação do ObjetoJogo */
-int ObjetoJogo::get_valor_hidratacao(){
+float ObjetoJogo::get_valor_hidratacao(){
 	return valor_hidratacao;
 }
 
 /*! Define valor de hidratação do ObjetoJogo */
-void ObjetoJogo::set_valor_hidratacao(int novo_valor_hidratacao){
+void ObjetoJogo::set_valor_hidratacao(float novo_valor_hidratacao){
 	valor_hidratacao = novo_valor_hidratacao;
 }
